feat: console now works!
